+++ packages/expo-router/src/toolbar/elements.tsx
@@ -1,9 +1,10 @@
 import { Children, isValidElement, useId, type ReactNode } from 'react';
 import { Platform, StyleSheet, type ColorValue, type StyleProp } from 'react-native';
-import { SearchBar, type SearchBarProps } from 'react-native-screens';
+import { type SearchBarProps } from 'react-native-screens';
 import type { SFSymbol } from 'sf-symbols-typescript';
 
 import { RouterToolbarHost, RouterToolbarItem } from './native';
+import Stack from '../layouts/StackClient';
 import { InternalLinkPreviewContext } from '../link/InternalLinkPreviewContext';
 import { LinkMenuAction, type LinkMenuActionProps } from '../link/elements';
 import { NativeLinkPreviewAction } from '../link/preview/native';
@@ -416,7 +417,9 @@ export interface ToolbarSearchBarProps extends SearchBarProps {
  * A search bar component for the toolbar.
  * It should only be used as a child of `Toolbar`.
  *
- * > **Note**: Once Toolbar.SearchBar is used, the header search bar cannot be used for that screen.
+ * > **Note**: The stack header must be visible in order to add a search bar.
+ * > Once `Toolbar.SearchBar` is used, the search bar will no longer be available
+ * > in this screen’s header.
  *
  * @example
  * ```tsx
@@ -427,7 +430,7 @@ export interface ToolbarSearchBarProps extends SearchBarProps {
  * </Toolbar>
  * ```
  *
- * @see [Apple documentation](https://developer.apple.com/documentation/uikit/uinavigationitem/searchbarplacementbarbuttonitem) for more information.
+ * @see [React Native Screens SearchBarProps](https://github.com/software-mansion/react-native-screens/blob/main/src/types.tsx#L822) for more information about available props.
  *
  * @platform ios 26+
  */
@@ -454,7 +457,7 @@ export const ToolbarSearchBar = ({
       identifier={id}
       sharesBackground={sharesBackground}
       type="searchBar">
-      <SearchBar {...searchBarProps} placement={placement} />
+      <Stack.Screen options={{ headerSearchBarOptions: { ...searchBarProps, placement } }} />
     </RouterToolbarItem>
   );
 };